fp64: Refactor and export 64-bit constants
The 80 bit float handling used to do a fair amount of 64-bit float
manipulation for conversions. This meant that potentially useful
64-bit float constants were not exported.

This changeset refactors the code slightly to move 64-bit float
constants to a separate file (fputils/fp64.h) and the library's
representation of floats (both 64-bit and 80-bit floats) to a new file
(fputils/fptypes.h).