Update license to match gem5
This change aligns the license from a 2-clause BSD license to a
3-clause BSD license to match gem5's license. It also changes the
license headers to have the same formatting as gem5.

This library has only seen one commit by a different author (Anthony
Gutierrez):

043765 fputils: add explicit padding to correctly align fp80_t

Tony was consulted on the 18th of October and acknowledged that he was
fine with the license change.